COOP: emit report whenever there is another window in the bcg
This CL changes reporting of browsing context group switches for COOP.
Prior to this CL, a report would only be generated if the frame had an
opener at the start of navigation. Following this CL, the report is
generated whenever there is another window in the browsing context
group which is in line with the spec.

Bug: 1148709
Change-Id: If2bf80365896a50692b4c02338a3fedf27348df7